Move FunctionalTestHelper to util
