+++ code/edge/build.gradle
@@ -246,15 +246,11 @@ dependencies {
     testImplementation 'org.json:json:20180813'
 
     // TODO - Uncomment next lines when 2.0 is published to Maven
-    implementation fileTree(dir: 'libs', include: ['edgeidentity-release-2.0.0.aar', 'edgeconsent-release-2.0.0.aar'])
 //    androidTestImplementation 'com.adobe.marketing.mobile:edgeidentity:1.+'
 //    androidTestImplementation 'com.adobe.marketing.mobile:edgeconsent:1.+'
+    implementation fileTree(dir: 'libs', include: ['edgeidentity-release-2.0.0.aar', 'edgeconsent-release-2.0.0.aar'])
 
     androidTestImplementation "androidx.test.ext:junit:${rootProject.ext.junitVersion}"
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
-    // Using older version of PowerMock as it supports minimum Android API < 26
-    // the root issue appears to be a dependency with Objenesis in Mockito-Core
-    // where Objenesis 2 requires minimum Android API 26
-    //androidTestImplementation 'org.powermock:powermock-module-junit4:1+'
     androidTestImplementation 'com.fasterxml.jackson.core:jackson-databind:2.12.7'
 }
