+++ code/edge/src/androidTest/java/com/adobe/marketing/mobile/SampleFunctionalTests.java
@@ -20,6 +20,7 @@ import com.adobe.marketing.mobile.edge.identity.Identity;
 import com.adobe.marketing.mobile.services.HttpConnecting;
 import com.adobe.marketing.mobile.services.TestableNetworkRequest;
 import com.adobe.marketing.mobile.util.FunctionalTestConstants;
+import com.adobe.marketing.mobile.util.FunctionalTestHelper;
 import com.adobe.marketing.mobile.util.FunctionalTestUtils;
 import java.util.Arrays;
 import java.util.HashMap;
